# web/sui/openapi.yaml
openapi: 3.0.0
info:
  title: 'ZSF API'
  version: '1.1'
servers:
  -
    url: /api/1.1
    description: 'ZSF API'
paths:
...
  '/Docs/Admin/{resource}':
    get:
      tags:
        - DocsAdmin
      summary: 'List rows of a documentation table'
      description: 'Every row, including ones removed or not yet introduced at any version. Filterable by parent (moduleId, symbolId, pageId, versionId, courseId, lessonId), mode, or kind where applicable.'
      operationId: docsAdminList
      parameters:
        -
          name: resource
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DocAdminResource'
        -
          name: moduleId
          in: query
          required: false
          schema:
            type: integer
        -
          name: symbolId
          in: query
          required: false
          schema:
            type: integer
        -
          name: pageId
          in: query
          required: false
          schema:
            type: integer
        -
          name: versionId
          in: query
          required: false
          schema:
            type: integer
        -
          name: courseId
          in: query
          required: false
          schema:
            type: integer
        -
          name: lessonId
          in: query
          required: false
          schema:
            type: integer
        -
          name: mode
          in: query
          required: false
          schema:
            type: string
        -
          name: kind
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocAdminRecord'
        '404':
          description: 'Unknown resource'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
    post:
      tags:
        - DocsAdmin
      summary: 'Create a row'
      description: "Body carries the row's fields in camelCase. Validation failures (duplicates, overlapping contract ranges, unknown versions) return 400 with a message."
      operationId: docsAdminCreate
      parameters:
        -
          name: resource
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DocAdminResource'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocAdminRecord'
      responses:
        '200':
          description: 'Created row'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocAdminRecord'
        '400':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  '/Docs/Admin/{resource}/{id}':
    get:
      tags:
        - DocsAdmin
      summary: 'Get a row'
      description: 'Retrieves one row.'
      operationId: docsAdminGet
      parameters:
        -
          name: resource
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DocAdminResource'
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocAdminRecord'
        '404':
          description: 'Not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
    put:
      tags:
        - DocsAdmin
      summary: 'Update a row'
      description: 'Only fields present in the body change.'
      operationId: docsAdminUpdate
      parameters:
        -
          name: resource
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DocAdminResource'
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocAdminRecord'
      responses:
        '200':
          description: 'Updated row'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocAdminRecord'
        '400':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
        '404':
          description: 'Not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
    delete:
      tags:
        - DocsAdmin
      summary: 'Delete a row'
      description: "Children cascade (a module's symbols, a page's samples). Rows still referenced elsewhere, such as a version with contracts, are rejected."
      operationId: docsAdminDelete
      parameters:
        -
          name: resource
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DocAdminResource'
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                properties:
                  deleted: { type: integer }
                type: object
        '400':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
        '404':
          description: 'Not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  '/Docs/Admin/{resource}/Reorder':
    post:
      tags:
        - DocsAdmin
      summary: 'Reorder rows'
      description: 'Sets sortOrder (modules, symbols, changes, courses, context options) or ordinal (lessons, steps) to the position of each id in the list, 1-based.'
      operationId: docsAdminReorder
      parameters:
        -
          name: resource
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DocAdminResource'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - ids
              properties:
                ids:
                  type: array
                  items: { type: integer }
              type: object
      responses:
        '200':
          description: 'Rows in their new order'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocAdminRecord'
        '400':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  '/Docs/Admin/Versions/{id}/Latest':
    post:
      tags:
        - DocsAdmin
      summary: 'Mark a version as latest'
      description: 'Marks a version as the one `latest` resolves to.'
      operationId: docsAdminMarkLatest
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'The version'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocAdminRecord'
        '404':
          description: 'Not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  '/Docs/Admin/Pages/{id}/Symbols':
    get:
      tags:
        - DocsAdmin
      summary: 'Symbols linked to a page'
      description: 'The symbols linked to a page.'
      operationId: docsAdminGetPageSymbols
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocAdminSymbolLink'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
    put:
      tags:
        - DocsAdmin
      summary: "Replace a page's symbol links"
      description: 'Body is the full list; each entry needs symbolId or ref, and a role (subject or mentions).'
      operationId: docsAdminSetPageSymbols
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/DocAdminSymbolLink'
      responses:
        '200':
          description: 'The new links'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocAdminSymbolLink'
        '400':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  '/Docs/Admin/Changes/{id}/Symbols':
    get:
      tags:
        - DocsAdmin
      summary: 'Symbols a change affects'
      description: 'The symbols a change affects.'
      operationId: docsAdminGetChangeSymbols
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocAdminSymbolLink'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
    put:
      tags:
        - DocsAdmin
      summary: "Replace a change's affected symbols"
      description: 'Body is the full list of symbol ids, refs, or objects with symbolId / ref.'
      operationId: docsAdminSetChangeSymbols
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '200':
          description: 'The new links'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocAdminSymbolLink'
        '400':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocError'
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
...
  /Docs/Admin/Export:
    get:
      tags:
        - DocsAdmin
      summary: 'Export the whole documentation set as a bundle'
      description: 'Exports every documentation record as a bundle.'
      operationId: docsAdminExport
      responses:
        '200':
          description: Bundle
          content:
            application/json:
              schema:
                type: object
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
...
components:
  schemas:
...
    DocError:
      description: 'Error responses carry a plain message string'
      type: string
      example: "Unknown version 'v9.9'"
    DocAdminResource:
      description: 'Documentation tables exposed for authoring'
      type: string
      enum:
        - Versions
        - Modules
        - Symbols
        - Contracts
        - Pages
        - Samples
        - Changes
        - Courses
        - Lessons
        - Steps
        - ContextOptions
    DocAdminRecord:
      description: 'One row of a documentation table with camelCase field names; `id` is assigned by the server. Contracts carry decoded params/returns/throws; symbols carry their `ref`.'
      properties:
        id:
          type: integer
      type: object
      additionalProperties: true
    DocAdminSymbolLink:
      properties:
        symbolId:
          type: integer
        ref:
          type: string
          example: 'tessel/query#QueryOptions.staleTime'
        name:
          type: string
        kind:
          type: string
        role:
          description: 'Pages only'
          type: string
          enum:
            - subject
            - mentions
      type: object
...
  securitySchemes:
    header_token:
      type: http
      scheme: bearer
    cookie_token:
      type: apiKey
      name: zsf_token
      in: cookie
    admin_header_token:
      type: http
      scheme: bearer
    admin_cookie_token:
      type: apiKey
      name: zsf_token
      in: cookie
...